Move the Dodo::Context class to its own file.
At the moment the `Context` class shares a file with the `Runner` class.  To avoid overly large files it should be moved to its own file.
