+++ lib/dodo.rb
@@ -2,6 +2,7 @@
 
 require 'dodo/serial_timeline'
 require 'dodo/moment'
+require 'dodo/scope'
 require 'dodo/runner'
 require 'dodo/parallel_timeline'
 require 'dodo/progress'
